@@ -11,45 +11,50 @@
     "build": "ng build --prod",
     "test": "ng test",
     "lint": "ng lint",
+    "build:lib": "rimraf dist && ng-packagr -p saturn-datepicker/package.json",
     "packagr": "ng-packagr -p ng-package.json"
   },
   "private": false,
   "dependencies": {},
   "devDependencies": {
-    "@angular/animations": "^5.2.4",
-    "@angular/common": "^5.2.4",
-    "@angular/cdk": "^5.0.3",
-    "@angular/compiler": "^5.2.4",
-    "@angular/compiler-cli": "^5.2.4",
-    "@angular/material": "^5.2.4",
-    "@angular/core": "^5.2.4",
-    "@angular/forms": "^5.2.4",
-    "@angular/http": "^5.2.4",
-    "@angular/platform-browser": "^5.2.4",
-    "@angular/platform-browser-dynamic": "^5.2.4",
-    "@angular/router": "^5.2.4",
-    "@angular/cli": "1.7.4",
+    "@angular/common": "6.0.0-rc.6",
+    "@angular/compiler": "6.0.0-rc.6",
+    "@angular/core": "6.0.0-rc.6",
+    "@angular/forms": "6.0.0-rc.6",
+    "@angular/http": "6.0.0-rc.6",
+    "@angular/platform-browser": "6.0.0-rc.6",
+    "@angular/platform-browser-dynamic": "6.0.0-rc.6",
+    "@angular/router": "6.0.0-rc.6",
     "core-js": "^2.4.1",
-    "rxjs": "^5.5.7",
+    "rxjs": "6.0.0",
+    "rxjs-compat": "^6.0.0",
     "systemjs": "0.19.43",
-    "tslib": "^1.7.1",
-    "@types/jasmine": "~2.8.3",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
-    "jasmine-core": "~2.8.0",
-    "jasmine-spec-reporter": "~4.2.1",
+    "@angular-devkit/build-angular": "~0.5.0",
+    "@angular-devkit/schematics": "^0.5.0",
+    "@angular/cli": "6.0.0-rc.10",
+    "@angular/compiler-cli": "6.0.0-rc.6",
+    "@angular/language-service": "6.0.0-rc.6",
+    "@angular/material": "6.0.0-rc.14",
+    "@angular/cdk": "6.0.0-rc.14",
+    "@types/jasmine": "~2.8.0",
+    "@types/node": "~9.6.0",
+    "codelyzer": "~4.3.0",
+    "cpr": "^3.0.0",
+    "jasmine-core": "~3.1.0",
+    "jasmine-spec-reporter": "~4.2.0",
     "karma": "~2.0.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
+    "karma-cli": "~1.0.1",
+    "karma-coverage-istanbul-reporter": "^1.3.0",
     "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "ng-packagr": "^2.4.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "tsickle": "^0.27.2",
-    "typescript": "~2.7.2",
-    "zone.js": "^0.8.19"
+    "karma-jasmine-html-reporter": "^1.0.0",
+    "ng-packagr": "3.0.0-rc.2",
+    "protractor": "~5.3.0",
+    "rimraf": "^2.6.1",
+    "ts-node": "~6.0.0",
+    "tsickle": "^0.27.3",
+    "tslib": "^1.7.1",
+    "tslint": "~5.9.0",
+    "typescript": "~2.7.2"
   }
 }
