@@ -6,16 +6,27 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {InjectionToken, LOCALE_ID} from '@angular/core';
-import {Observable} from 'rxjs/Observable';
-import {Subject} from 'rxjs/Subject';
+import {inject, InjectionToken, LOCALE_ID} from '@angular/core';
+import {Observable, Subject} from 'rxjs';
 
 /** InjectionToken for datepicker that can be used to override default locale code. */
-export const MAT_DATE_LOCALE = new InjectionToken<string>('MAT_DATE_LOCALE');
+export const MAT_DATE_LOCALE = new InjectionToken<string>('MAT_DATE_LOCALE', {
+  providedIn: 'root',
+  factory: MAT_DATE_LOCALE_FACTORY,
+});
 
-/** Provider for MAT_DATE_LOCALE injection token. */
+/** @docs-private */
+export function MAT_DATE_LOCALE_FACTORY(): string {
+  return inject(LOCALE_ID);
+}
+
+/**
+ * No longer needed since MAT_DATE_LOCALE has been changed to a scoped injectable.
+ * If you are importing and providing this in your code you can simply remove it.
+ * @deprecated
+ * @deletion-target 7.0.0
+ */
 export const MAT_DATE_LOCALE_PROVIDER = {provide: MAT_DATE_LOCALE, useExisting: LOCALE_ID};
-
 
 /** Adapts type `D` to be usable as a date by cdk-based components that work with dates. */
 export abstract class DateAdapter<D> {
@@ -196,7 +207,7 @@ export abstract class DateAdapter<D> {
    * Attempts to deserialize a value to a valid date object. This is different from parsing in that
    * deserialize should only accept non-ambiguous, locale-independent formats (e.g. a ISO 8601
    * string). The default implementation does not allow any deserialization, it simply checks that
-   * the given value is already a valid date object or null. The `<mat-datepicker>` will call this
+   * the given value is already a valid date object or null. The `<sat-datepicker>` will call this
    * method on all of it's `@Input()` properties that accept dates. It is therefore possible to
    * support passing values from your backend directly to these properties by overriding this method
    * to also deserialize the format used by your backend.
