@@ -7,19 +7,33 @@
  */
 
 import {
-  AfterContentInit,
-  ChangeDetectionStrategy,
-  ChangeDetectorRef,
-  Component,
-  EventEmitter,
-  Inject,
-  Input,
-  Optional,
-  Output,
-  ViewEncapsulation,
+    DOWN_ARROW,
+    END,
+    ENTER,
+    HOME,
+    LEFT_ARROW,
+    PAGE_DOWN,
+    PAGE_UP,
+    RIGHT_ARROW,
+    UP_ARROW,
+} from '@angular/cdk/keycodes';
+import {
+    AfterContentInit,
+    ChangeDetectionStrategy,
+    ChangeDetectorRef,
+    Component,
+    EventEmitter,
+    Inject,
+    Input,
+    Optional,
+    Output,
+    ViewChild,
+    ViewEncapsulation,
 } from '@angular/core';
-import {DateAdapter, MAT_DATE_FORMATS, MatDateFormats} from '../datetime';
-import {SatCalendarCell} from './calendar-body';
+import {DateAdapter} from '../datetime/date-adapter';
+import {MAT_DATE_FORMATS, MatDateFormats} from '../datetime/date-formats';
+import {Directionality} from '@angular/cdk/bidi';
+import {SatCalendarBody, SatCalendarCell} from './calendar-body';
 import {createMissingDateImplError} from './datepicker-errors';
 
 
@@ -36,8 +50,7 @@ const DAYS_PER_WEEK = 7;
   templateUrl: 'month-view.html',
   exportAs: 'matMonthView',
   encapsulation: ViewEncapsulation.None,
-  preserveWhitespaces: false,
-  changeDetection: ChangeDetectionStrategy.OnPush,
+  changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class SatMonthView<D> implements AfterContentInit {
 
@@ -80,9 +93,10 @@ export class SatMonthView<D> implements AfterContentInit {
   @Input()
   get activeDate(): D { return this._activeDate; }
   set activeDate(value: D) {
-    let oldActiveDate = this._activeDate;
-    this._activeDate =
+    const oldActiveDate = this._activeDate;
+    const validDate =
         this._getValidDateOrNull(this._dateAdapter.deserialize(value)) || this._dateAdapter.today();
+    this._activeDate = this._dateAdapter.clampDate(validDate, this.minDate, this.maxDate);
     if (!this._hasSameMonthAndYear(oldActiveDate, this._activeDate)) {
       this._init();
     }
@@ -123,6 +137,12 @@ export class SatMonthView<D> implements AfterContentInit {
   /** Emits when any date is selected. */
   @Output() readonly _userSelection: EventEmitter<void> = new EventEmitter<void>();
 
+  /** Emits when any date is activated. */
+  @Output() readonly activeDateChange: EventEmitter<D> = new EventEmitter<D>();
+
+  /** The body of calendar table */
+  @ViewChild(SatCalendarBody) _matCalendarBody;
+
   /** The label for this month (e.g. "January 2017"). */
   _monthLabel: string;
 
@@ -144,9 +164,10 @@ export class SatMonthView<D> implements AfterContentInit {
   /** The names of the weekdays. */
   _weekdays: {long: string, narrow: string}[];
 
-  constructor(@Optional() public _dateAdapter: DateAdapter<D>,
+  constructor(private _changeDetectorRef: ChangeDetectorRef,
               @Optional() @Inject(MAT_DATE_FORMATS) private _dateFormats: MatDateFormats,
-              private _changeDetectorRef: ChangeDetectorRef) {
+              @Optional() public _dateAdapter: DateAdapter<D>,
+              @Optional() private _dir?: Directionality) {
     if (!this._dateAdapter) {
       throw createMissingDateImplError('DateAdapter');
     }
@@ -169,6 +190,7 @@ export class SatMonthView<D> implements AfterContentInit {
 
   ngAfterContentInit() {
     this._init();
+    this._focusActiveCell();
   }
 
   /** Handles when a new date is selected. */
@@ -192,9 +214,73 @@ export class SatMonthView<D> implements AfterContentInit {
       const selectedYear = this._dateAdapter.getYear(this.activeDate);
       const selectedMonth = this._dateAdapter.getMonth(this.activeDate);
       const selectedDate = this._dateAdapter.createDate(selectedYear, selectedMonth, date);
+
       this.selectedChange.emit(selectedDate);
       this._userSelection.emit();
     }
+  }
+
+  /** Handles keydown events on the calendar body when calendar is in month view. */
+  _handleCalendarBodyKeydown(event: KeyboardEvent): void {
+    // TODO(mmalerba): We currently allow keyboard navigation to disabled dates, but just prevent
+    // disabled ones from being selected. This may not be ideal, we should look into whether
+    // navigation should skip over disabled dates, and if so, how to implement that efficiently.
+
+    const oldActiveDate = this._activeDate;
+
+    const isRtl = this._isRtl();
+    switch (event.keyCode) {
+      case LEFT_ARROW:
+        this.activeDate = this._dateAdapter.addCalendarDays(this._activeDate, isRtl ? 1 : -1);
+        break;
+      case RIGHT_ARROW:
+        this.activeDate = this._dateAdapter.addCalendarDays(this._activeDate, isRtl ? -1 : 1);
+        break;
+      case UP_ARROW:
+        this.activeDate = this._dateAdapter.addCalendarDays(this._activeDate, -7);
+        break;
+      case DOWN_ARROW:
+        this.activeDate = this._dateAdapter.addCalendarDays(this._activeDate, 7);
+        break;
+      case HOME:
+        this.activeDate = this._dateAdapter.addCalendarDays(this._activeDate,
+            1 - this._dateAdapter.getDate(this._activeDate));
+        break;
+      case END:
+        this.activeDate = this._dateAdapter.addCalendarDays(this._activeDate,
+            (this._dateAdapter.getNumDaysInMonth(this._activeDate) -
+              this._dateAdapter.getDate(this._activeDate)));
+        break;
+      case PAGE_UP:
+        this.activeDate = event.altKey ?
+            this._dateAdapter.addCalendarYears(this._activeDate, -1) :
+            this._dateAdapter.addCalendarMonths(this._activeDate, -1);
+        break;
+      case PAGE_DOWN:
+        this.activeDate = event.altKey ?
+            this._dateAdapter.addCalendarYears(this._activeDate, 1) :
+            this._dateAdapter.addCalendarMonths(this._activeDate, 1);
+        break;
+      case ENTER:
+        if (!this.dateFilter || this.dateFilter(this._activeDate)) {
+          this._dateSelected(this._dateAdapter.getDate(this._activeDate));
+          this._userSelection.emit();
+          // Prevent unexpected default actions such as form submission.
+          event.preventDefault();
+        }
+        return;
+      default:
+        // Don't prevent default or focus active cell on keys that we don't explicitly handle.
+        return;
+    }
+
+    if (this._dateAdapter.compareDate(oldActiveDate, this.activeDate)) {
+      this.activeDateChange.emit(this.activeDate);
+    }
+
+    this._focusActiveCell();
+    // Prevent unexpected default actions such as form submission.
+    event.preventDefault();
   }
 
   /** Initializes this month view. */
@@ -214,6 +300,11 @@ export class SatMonthView<D> implements AfterContentInit {
 
     this._createWeekCells();
     this._changeDetectorRef.markForCheck();
+  }
+
+  /** Focuses the active cell after the microtask queue is empty. */
+  private _focusActiveCell() {
+    this._matCalendarBody._focusActiveCell();
   }
 
   /** Creates SatCalendarCells for the dates in this month. */
@@ -266,6 +357,11 @@ export class SatMonthView<D> implements AfterContentInit {
   private _getValidDateOrNull(obj: any): D | null {
     return (this._dateAdapter.isDateInstance(obj) && this._dateAdapter.isValid(obj)) ? obj : null;
   }
+
+    /** Determines whether the user has the RTL layout direction. */
+    private _isRtl() {
+        return this._dir && this._dir.value === 'rtl';
+    }
   /** Updates range full parameter on each begin or end of interval update.
    * Necessary to display calendar-body correctly
    */
